Chore: Migrate the Alternative class to the main language_step.py. This ensures that we only use one file, but also has option to use other class when performance of using Flow class is bad.
